deployment files/requirements.txt
# Kaluwala CSR Libraries - Production Dependencies

# Core Flask
Flask>=3.0.0
Werkzeug>=3.0.0

# Database
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0
psycopg2-binary>=2.9.0

# Authentication
Flask-Login>=0.6.0

# Email
Flask-Mail>=0.10.0

# PDF Generation
reportlab>=4.0.0

# Environment
python-dotenv>=1.0.0

# Images
Pillow>=10.0.0

# Scheduler
APScheduler>=3.10.0

# Charts
matplotlib>=3.8.0

# Excel
openpyxl>=3.1.0

# Utils
email-validator>=2.1.0

# Production Server
gunicorn>=21.2.0

# Migrations
alembic>=1.13.0
